requirements.txt
alembic~=1.15.2
asyncpg~=0.30.0
bcrypt~=4.3.0
email-validator~=2.2.0
fastapi~=0.115.11
fastapi-pagination~=0.13.3
numpy~=2.2.4
openpyxl~=3.1.5
pandas~=2.2.3
psycopg2~=2.9.10
pydantic~=2.11.3
python-dotenv~=1.1.0
setuptools~=70.0.0
SQLAlchemy~=2.0.40
sqlmodel~=0.0.24
uvicorn~=0.34.2
